+++ Runtime/wechat-default/wx-game-kit/custom/index.js
@@ -1,0 +1,3 @@
+import './custom-instance';
+import './handlers';
+import './bind-events';