+++ Runtime/wechat-default/wx-game-kit/loader-config.js
@@ -1,0 +1,2 @@
+const loaderConfig = JSON.parse(wx.getFileSystemManager().readFileSync('wx-game-kit/config.json', 'utf-8'));
+export default loaderConfig;