+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Others - 其他
 * 普通：EmscriptenGLX 微信压缩纹理占位图改成透明图
 * 普通：修复多线程下无法transferControlToOffscreen的问题
 
-## 2026-8-5 v0.1.33 【普通更新】
+## 2026-8-25 v0.1.33 【普通更新】
 ### Feature
 * 普通：升级基础库版本
 * 普通：线程支持相关优化
@@ -21,6 +21,7 @@ Others - 其他
 * 普通：EmscriptenGLX支持多线程
 * 普通：EmscriptenGLX 支持压缩纹理
 * 普通：优化 EmscriptenGLX shader warmup 速度
+* 普通：优化 EmscriptenGLX 对象 id 生成机制，避免宿主多 gl context id 冲突导致渲染异常
 * 普通：优化转换面板的 iOSMetal 提示语
 * 普通：鸿蒙视频音频播放适配
 ### Fixed
@@ -28,6 +29,8 @@ Others - 其他
 * 普通：修复语音房和webaudio的冲突问题
 * 普通：兼容新版启动插件isErasableFile回调的relativePath，避免AB文件被误清理
 * 普通：iOS平台屏蔽getCommonFont拉取系统字体能力
+* 普通：补全 WX_SyncFunction_tnnt 桥接实现
+* 普通：修复预下载缓存失效问题
 
 ## 2026-3-16 v0.1.32 【普通更新】
 ### Feature